requirements.txt
# ARIA - Adaptive Retrieval with Intelligent Anchoring
# Requirements for production deployment

# Core dependencies
numpy>=1.21.0
scikit-learn>=1.0.0
PyYAML>=6.0

# Optional: Semantic search (recommended for better results)
sentence-transformers>=2.2.0

# Optional: PDF support
PyPDF2>=3.0.0

# Optional: Word document support  
python-docx>=0.8.11

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# mypy>=0.990
